requirements.txt
python-dotenv
tensorflow==2.12.0
opencv-python==4.7.0.72
numpy==1.23.5
matplotlib==3.7.1
